Rename the FakeClock to "Clock"
